calorie-tracker/model/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
datasets>=2.15.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
accelerate>=0.24.0
evaluate>=0.4.0
onnx>=1.14.0
onnxruntime>=1.16.0
huggingface-hub>=0.19.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
